refactor(runner): extract Terminal/FileExplorer/Git into ServiceHandler pattern
- Add ServiceHandler interface, ServiceRegistry, ServiceEnvelope types
- Extract TerminalService (new_terminal, terminal_input, terminal_resize, kill_terminal, list_terminals)
- Extract FileExplorerService (list_files, search_files, read_file)
- Extract GitService (git_status, git_diff)
- Wire services via registry in daemon.ts
- daemon.ts shrinks from 1253 → 892 lines
- No behavior changes — relay protocol unchanged
